refactor: view로직 domain에서 제거
