+++ src/main/java/lotto/view/input/LottoBuyInputView.java
@@ -29,7 +29,7 @@ public class LottoBuyInputView extends Input{
     }
 
     private void printRequestLottoPurchaseAmount() {
-        System.out.println(LottoBuyMessage.PURCHASE_AMOUNT_PROMPT_MESSAGE);
+        System.out.println(LottoBuyMessage.PURCHASE_AMOUNT_PROMPT_MESSAGE.getMassage());
     }
 
     private void validateLottoAmount(int amount, int denomination) {
@@ -48,7 +48,7 @@ public class LottoBuyInputView extends Input{
     }
 
     private void printRequestWinningLottoNumbers() {
-        System.out.println(LottoBuyMessage.WINNING_NUMBERS_PROMPT_MESSAGE);
+        System.out.println(LottoBuyMessage.WINNING_NUMBERS_PROMPT_MESSAGE.getMassage());
     }
 
     private List<Integer> getNumbers() {
@@ -72,7 +72,7 @@ public class LottoBuyInputView extends Input{
     }
 
     private void printRequestBonusLottoNumber() {
-        System.out.println(LottoBuyMessage.BONUS_NUMBER_PROMPT_MESSAGE);
+        System.out.println(LottoBuyMessage.BONUS_NUMBER_PROMPT_MESSAGE.getMassage());
     }
 
     private void validateBonusLottoNumber(int bonusNumber) {
